fix: pass scripts through the ViewModel instead of navigation arguments
Fixes the Crashlytics issue: IllegalArgumentException "Navigation
destination that matches route ...AppRoutes.Terminal?script=..." in
RootGraph (2.0 (15), Android 16).

- Terminal/Editor routes now carry no script; MainViewModel.openScript()
  hands it over and each destination reads it once per back stack entry
- Every navigation writes the slot first, so there is no stale value
- Running passes the editor's live buffer, so edits are included
- Drop MicroScript.asJson, String.asMicroScript() and @Serializable
- Move AppRoutes to the model package
